.tsx files are kebab-cased
