+++ packages/wallet/src/__stories__/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import WalletContainer from '../containers/Wallet';
+import WalletContainer from '../containers/wallet';
 import { Provider } from 'react-redux';
 import * as walletStates from '../redux/state';
 import * as channelStates from '../redux/channel-state/state';
@@ -8,7 +8,7 @@ import * as fundingStates from '../redux/funding-state/state';
 import '../index.scss';
 import * as scenarios from '../redux/__tests__/test-scenarios';
 import { bigNumberify } from 'ethers/utils';
-import NetworkStatus from '../components/NetworkStatus';
+import NetworkStatus from '../components/network-status';
 import { channelFunded } from '../redux/funding-state/state';
 import { EMPTY_OUTBOX_STATE } from '../redux/outbox/state';
