+++ packages/wallet/src/components/closing/closing-step.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faCheckCircle, faSpinner } from '@fortawesome/free-solid-svg-icons';
 import { faCircle } from '@fortawesome/free-regular-svg-icons';
-import SidebarLayout from '../SidebarLayout';
+import SidebarLayout from '../sidebar-layout';
 
 interface Props {
   step: number;
@@ -34,39 +34,7 @@ const icon = (iconStep: number, currentStep: number) => {
   }
 };
 
-// NOTE: the appearance of this modal is largely influenced by the amount of text in each message. Until a more robust front-end comes along, try to keep messages of the same length within each case block below.
-const message = (iconStep: number, currentStep: number) => {
-  switch (iconStep) {
-    case 1:
-      if (currentStep < iconStep) {
-        return 'Prepare your transaction';
-      } else if (currentStep === iconStep) {
-        return 'Preparing your transaction...';
-      } else {
-        return 'Withdrawal transaction sent';
-      }
-    case 2:
-      if (currentStep < iconStep) {
-        return 'Transaction will be confirmed';
-      } else if (currentStep === iconStep) {
-        return 'Waiting for confirmation...';
-      } else {
-        return 'Transaction confirmed';
-      }
-    case 3:
-      if (currentStep < iconStep) {
-        return 'Withdrawal will be confirmed';
-      } else if (currentStep === iconStep) {
-        return 'Waiting for confirmation...';
-      } else {
-        return 'Withdrawal successful!';
-      }
-    default:
-      return '';
-  }
-};
-
-export class WithdrawingStep extends React.PureComponent<Props> {
+export class ClosingStep extends React.PureComponent<Props> {
   render() {
     const currentStep = this.props.step;
     const children = this.props.children;
@@ -77,15 +45,15 @@ export class WithdrawingStep extends React.PureComponent<Props> {
         <ul className="fa-ul">
           <li style={{ padding: '0.7em 1em' }}>
             {icon(1, currentStep)}
-            {message(1, currentStep)}
+            Preparing your close transaction
           </li>
           <li style={{ padding: '0.7em 1em' }}>
             {icon(2, currentStep)}
-            {message(2, currentStep)}
+            Wait for confirmation
           </li>
           <li style={{ padding: '0.7em 1em' }}>
             {icon(3, currentStep)}
-            {message(3, currentStep)}
+            Close and Withdrawal Successful!
           </li>
         </ul>
         <div className="pb-2">{children}</div>
