+++ packages/wallet/src/components/funding/funding-error.tsx
