+++ packages/wallet/src/components/network-status.tsx
