+++ packages/wallet/src/containers/Funding.tsx
@@ -6,8 +6,8 @@ import * as channelStates from '../redux/channel-state/state';
 import * as actions from '../redux/actions';
 
 import { unreachable } from '../utils/reducer-utils';
-import DirectFunding from './DirectFunding';
-import { FundingStep, Step } from '../components/funding/FundingStep';
+import DirectFunding from './direct-funding';
+import { FundingStep, Step } from '../components/funding/funding-step';
 import { WalletProcedure } from '../redux/types';
 
 interface Props {
