+++ packages/wallet/tslint.json
@@ -39,6 +39,6 @@
       "allow-leading-underscore"
     ],
     "no-switch-case-fall-through": true,
-    "file-name-casing": [true, { ".tsx": "pascal-case", ".ts": "kebab-case" }]
+    "file-name-casing": [true, "kebab-case"]
   }
 }
